Split supplier functionality into its own pagemaker
